% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{buildKdDataTree}
\alias{buildKdDataTree}
\title{Construct a KD-Tree using the data.tree package}
\usage{
buildKdDataTree(X, y, scale = TRUE, FUN = median, ...)
}
\arguments{
\item{X}{Data table of values on which to perform splits}

\item{y}{Vector with data that should be stored in the leafs}

\item{scale}{Should data be scaled beforehand, defaults to true}

\item{FUN}{Optional function to use for determining position of split. Defaults to median as this gives a balanced tree}

\item{...}{Optional arguments to pass on to FUN}
}
\description{
This algorithm works similarly as buildKdTree but allows for more fancy operation such as child-parent traversal
}
\author{
Jurian Baas
}
