% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{buildKdTree}
\alias{buildKdTree}
\title{Construct a KD-Tree using only basic vanilla R functionality}
\usage{
buildKdTree(X, y, FUN = median, ...)
}
\arguments{
\item{X}{Data frame of values on which to perform splits}

\item{y}{Vector with data that should be stored in the leafs}

\item{FUN}{Optional function to use for determining position of split. Defaults to median as this gives a balanced tree}

\item{...}{Optional arguments to pass on to FUN}
}
\description{
This algorithm builds a KD-Tree. NB currently duplicate values can cause undefined behaviour!
}
\author{
Jurian Baas
}
