% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{composite}
\alias{composite}
\title{Composite of an n-dimensional point}
\usage{
composite(p)
}
\arguments{
\item{p}{Point to calculate composite of}
}
\value{
The composite of the given point
}
\description{
Calculate the associated point in the composite number space
}
\author{
Jurian Baas
}
