% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{plot.kd.data.tree}
\alias{plot.kd.data.tree}
\title{Plot a scatter plot or a tree representation of a two dimensional KD-Tree}
\usage{
\method{plot}{kd.data.tree}(x, ..., plotting.type = "scatter")
}
\arguments{
\item{x}{An object of type kd.data.tree}

\item{...}{Optional arguments to pass on}

\item{plotting.type}{One of c("scatter", "tree")}
}
\description{
If plotting.type = "scatter", plots a scatter plot of a two dimensional KD-Tree by drawing the coordinate data as points and adding lines where the data is split. Otherwise, if plotting.type = "tree", draws the two dimensional KD-Tree as a top down tree. Leafs are made square as per convention.
}
\author{
Jurian Baas
}
