% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{plot.kd.tree}
\alias{plot.kd.tree}
\title{Plot an object of type kd.tree}
\usage{
\method{plot}{kd.tree}(x, ...)
}
\arguments{
\item{x}{An object of type kd.tree to plot}

\item{...}{Optional arguments to pass on}
}
\description{
Not working ATM, please use plot.kd.data.tree instead
}
\author{
Jurian Baas
}
