% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdTree.R
\name{traverseKdTree}
\alias{traverseKdTree}
\title{Traverse an object of type kd.tree}
\usage{
traverseKdTree(kdtree, FUN, ...)
}
\arguments{
\item{kdtree}{An object of type kd.tree}

\item{FUN}{This function will be given a node as the first parameter}

\item{...}{Optional arguments to pass on to FUN}
}
\description{
Traverses the KD-Tree recursively, applying the function FUN to each node
}
\author{
Jurian Baas
}
