% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreperation.R
\name{combine.weather.datasets}
\alias{combine.weather.datasets}
\title{Combine multiple weather data sets with station information}
\usage{
combine.weather.datasets(station.file, weather.files, columns, tz = "UTC")
}
\arguments{
\item{station.file}{Path to file with station information, e.g. "data/stations.csv"}

\item{weather.files}{A character vector with paths of weather information, e.g. c("data/precipitation.csv", "data/temperature.csv")}

\item{columns}{Which columns to take from the merged result, e.g. c("longitude", "latitude", "datetime", "precipitation",  "temperature")}

\item{tz}{Optional timezone setting, defaults to "UTC"}
}
\value{
A datatable created by merging supplied data sets with each other and the station data. Datetime information is converted to unix timestamp
}
\description{
This function makes several assumptions about the structure of supplied datasets. Weather data is assumed to be of the form <station, datetime, value>. Station data is assumed to be minimally of the form <station, longitude, latitude>. This function was specifially designed to merge data taken from http://www.sciamachy-validation.org/climatology/daily_data/selection.cgi
}
\seealso{
http://www.sciamachy-validation.org/climatology/daily_data/selection.cgi
}
\author{
Jurian Baas
}
